Rename index.rs to recording_index.rs
