+++ crates/framework/src/lib.rs
@@ -2,22 +2,22 @@ mod additional_output;
 mod future_queue;
 mod historic_databases;
 mod historic_input;
-mod index;
 mod main_output;
 mod multiple_buffer;
 mod panic;
 mod parameters;
 mod perception_databases;
 mod perception_input;
+mod recording_index;
 
 pub use additional_output::{should_be_filled, AdditionalOutput};
 pub use future_queue::{future_queue, Consumer, Item, Producer, Update, Updates};
 pub use historic_databases::HistoricDatabases;
 pub use historic_input::HistoricInput;
-pub use index::{RecordingFrame, RecordingIndex};
 pub use main_output::MainOutput;
 pub use multiple_buffer::{multiple_buffer_with_slots, Reader, ReaderGuard, Writer, WriterGuard};
 pub use panic::deserialize_not_implemented;
 pub use parameters::Parameters;
 pub use perception_databases::PerceptionDatabases;
 pub use perception_input::PerceptionInput;
+pub use recording_index::{RecordingFrame, RecordingIndex};
