+++ crates/framework/src/recording_index.rs
